backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.200
openai==0.27.7
llama-index==0.6.9
pypdf==3.8.1
beautifulsoup4==4.12.2
requests==2.28.2
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
alembic==1.11.1
python-dotenv==1.0.0
pytest==7.3.1
httpx==0.24.0
python-magic==0.4.27
pytest-asyncio==0.21.0
pytest-cov==4.0.0
pytest-mock==3.10.0
